clWAVDataProvider: Clang compilation fix

+++ src/Decoders/WAV/WAVDataProvider.cpp
@@ -52,7 +52,7 @@ clWAVDataProvider::clWAVDataProvider( const std::shared_ptr<clBlob>& Data )
 			m_Format.m_SamplesPerSecond = Header->SampleRate;
 			m_Format.m_BitsPerSample    = Header->nBitsperSample;
 
-			m_DataSize = std::min( Header->DataSize, Data->GetDataSize() - sizeof(sWAVHeader) );
+			m_DataSize = std::min( static_cast<size_t>(Header->DataSize), Data->GetDataSize() - sizeof(sWAVHeader) );
 		}
 	}
 }
